# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API reto Torneo Solidario",
        "description": "API reto Torneo Solidario",
        "contact": {
            "email": "email@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/jugadores": {
            "get": {
                "tags": [
                    "jugadores"
                ],
                "summary": "Obtener todos los jugadores",
                "description": "Obtener todos los jugadores",
                "operationId": "getJugadores",
                "responses": {
                    "200": {
                        "description": "Lista de jugadores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/jugadores"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "jugadores"
                ],
                "summary": "Crear un jugador",
                "description": "Crear un nuevo jugador",
                "operationId": "postJugador",
                "requestBody": {
                    "description": "Datos del jugador",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "equipo_id",
                                    "nombre",
                                    "apellido1",
                                    "apellido2",
                                    "tipo",
                                    "estudio_id",
                                    "dni",
                                    "email",
                                    "telefono"
                                ],
                                "properties": {
                                    "equipo_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Jugador 1"
                                    },
                                    "apellido1": {
                                        "type": "string",
                                        "example": "Apellido 1"
                                    },
                                    "apellido2": {
                                        "type": "string",
                                        "example": "Apellido 2"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "jugador"
                                    },
                                    "estudio_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jugador@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jugador creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jugadores"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos no válidos"
                    }
                }
            }
        },
        "/api/jugadores/{id}": {
            "get": {
                "tags": [
                    "jugadores"
                ],
                "summary": "Obtener un jugador",
                "description": "Obtener un jugador por su id",
                "operationId": "getJugador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jugador encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jugadores"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "jugadores"
                ],
                "summary": "Actualizar un jugador",
                "description": "Actualizar un jugador por su id",
                "operationId": "updateJugador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del jugador",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "equipo_id",
                                    "nombre",
                                    "apellido1",
                                    "apellido2",
                                    "tipo",
                                    "estudio_id",
                                    "dni",
                                    "email",
                                    "telefono"
                                ],
                                "properties": {
                                    "equipo_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Jugador 1"
                                    },
                                    "apellido1": {
                                        "type": "string",
                                        "example": "Apellido 1"
                                    },
                                    "apellido2": {
                                        "type": "string",
                                        "example": "Apellido 2"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "jugador"
                                    },
                                    "estudio_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678A"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jugador@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jugador actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/jugadores"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "jugadores"
                ],
                "summary": "Eliminar un jugador",
                "description": "Eliminar un jugador por su id",
                "operationId": "deleteJugador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jugador eliminado"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "jugadores": {
                "title": "jugadores",
                "properties": {
                    "equipo_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "apellido1": {
                        "type": "string",
                        "example": "Perez"
                    },
                    "apellido2": {
                        "type": "string",
                        "example": "Gomez"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "jugador"
                    },
                    "estudio_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "dni": {
                        "type": "string",
                        "example": "12345678A"
                    },
                    "email": {
                        "type": "string",
                        "example": "juan@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "123456789"
                    },
                    "usuarioIdCreacion": {
                        "type": "integer",
                        "example": "1"
                    },
                    "fechaCreacion": {
                        "type": "timestamp",
                        "example": "2022-02-11 15:12:24"
                    },
                    "usuarioIdActualizacion": {
                        "type": "integer",
                        "example": "1"
                    },
                    "fechaActualizacion": {
                        "type": "timestamp",
                        "example": "2022-02-11 15:12:24"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}